Implemented the Score and Random System
The getScore(newScore) function gets the score from the game to the notif.jx. The scoreCounter gets the total score to print to getScore.

tempScore = newScore is for debugging.

This gets the Overall Score points of the whole game. Implemented with the random sequence number of Denzel where the score is the basis for when the notifications appear.